+++ locale/fr_FR.php
@@ -8,6 +8,8 @@
  */
 
 return [
+    //'"Delete User Data" will only delete the account data of the user, but NOT log them out if they are currently logged in, nor prevent the user from logging in again!' => '',
+    //'"Delete User" will only delete the account and associated data of the user, but NOT log the user out if they are currently logged in!' => '',
     '#Active Connections' => 'Nombre de connexions actives',
     //'ACL Permission List' => '',
     'Account' => 'Compte',
@@ -39,11 +41,13 @@ return [
     //'DNS Search Domain(s)' => '',
     //'DNS Server(s)' => '',
     //'DNS Suffix' => '',
+    //'Danger Zone' => '',
     'Date' => 'Date',
     'Date/Time' => 'Date/Heure',
     //'Default Gateway' => '',
     'Delete' => 'Supprimer',
     'Delete TOTP Secret' => 'Supprimer le secret TOTP',
+    //'Delete User' => '',
     //'Delete User Data' => '',
     'Details...' => 'Détails…',
     'Disable User' => 'Désactiver l\'utilisateur',